feat: move update check to dev
