+++ src/dev/mod.ts
@@ -1,3 +1,4 @@
+import { updateCheck } from "./update_check.ts";
 import {
   dirname,
   fromFileUrl,
@@ -10,6 +11,11 @@ import {
 import { error } from "./error.ts";
 
 const MIN_DENO_VERSION = "1.31.0";
+/**
+ * How often to query the server in ms.
+ * Right now that's every 24h
+ */
+const UPGRADE_CHECK_INTERVAL = 24 * 60 * 60 * 1000;
 
 export function ensureMinDenoVersion() {
   // Check that the minimum supported Deno version is being used.
@@ -163,6 +169,9 @@ export default manifest;
 
 export async function dev(base: string, entrypoint: string) {
   ensureMinDenoVersion();
+
+  // Run update check in background
+  updateCheck(UPGRADE_CHECK_INTERVAL).catch(() => {});
 
   entrypoint = new URL(entrypoint, base).href;
